move DnsQuestion struct into own module/file
