+++ src/protocol/mod.rs
@@ -2,10 +2,12 @@ pub mod domain;
 pub mod enums;
 pub mod header;
 pub mod packet;
+pub mod question;
 pub mod records;
 pub mod string;
 
 pub use enums::*;
 pub use header::*;
+pub use question::*;
 pub use records::*;
 pub use string::*;
